Drop KFunctionFromReferenceImpl, make FunctionImpl an interface
23 invokes in KFunctionFromReferenceImpl (and consequently, in FunctionImpl)
were needed before 1576160390aa9bc5d187477d53288dc686afa116: a wrapped function
reference must have had the necessary invoke to be called as an instance of a
specific function type. After 1576160390aa9bc5d187477d53288dc686afa116, this is
not needed anymore because KFunctionFromReferenceImpl is now an internal
implementation detail of reflection, and no invoke is ever called on it.